// app/shared/models/GrievancePriority.model.ts
export class GrievancePriorityModel {
     nGrievancePriorityId: number = 0;
     tGrievancePriority: string = null;
     nGrievancePriorityOrder: number = 0;
     bIsActive: boolean = false;
}
export class GrievancePriorityViewModel 
{
     nGrievanceTypeId: number = 0;
     tGrievancePriority: string="";
     nGrievancePriorityOrder: number = 0;
     dCreatedDt : string = ""
     bIsActive: boolean = false;
     nCreatedBy:number=0;
}